SQLError (char create/ logout) while using uint8. After switching to uint32 everything works fine.

+++ src/world/Player.cpp
@@ -2619,7 +2619,7 @@ void Player::SaveToDB(bool bNewCharacter /* =false */)
     
     ss << "', ";
 
-    ss << uint8(this->HasWonRbgToday());
+    ss << uint32(this->HasWonRbgToday());
 
     ss << ")";
